// gatsby/graphql-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type Internal = {
  content?: Maybe<Scalars['String']>;
  contentDigest: Scalars['String'];
  description?: Maybe<Scalars['String']>;
  fieldOwners?: Maybe<Array<Maybe<Scalars['String']>>>;
  ignoreType?: Maybe<Scalars['Boolean']>;
  mediaType?: Maybe<Scalars['String']>;
  owner: Scalars['String'];
  type: Scalars['String'];
};
// ...
export type Node = {
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type Node__Recipe = Node & {
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
  drupal_id?: Maybe<Scalars['String']>;
  drupal_internal__nid?: Maybe<Scalars['Int']>;
  drupal_internal__vid?: Maybe<Scalars['Int']>;
  langcode?: Maybe<Scalars['String']>;
  revision_timestamp?: Maybe<Scalars['Date']>;
  status?: Maybe<Scalars['Boolean']>;
  title?: Maybe<Scalars['String']>;
  created?: Maybe<Scalars['Date']>;
  changed?: Maybe<Scalars['Date']>;
  promote?: Maybe<Scalars['Boolean']>;
  sticky?: Maybe<Scalars['Boolean']>;
  default_langcode?: Maybe<Scalars['Boolean']>;
  revision_translation_affected?: Maybe<Scalars['Boolean']>;
  moderation_state?: Maybe<Scalars['String']>;
  path?: Maybe<Node__RecipePath>;
  content_translation_source?: Maybe<Scalars['String']>;
  content_translation_outdated?: Maybe<Scalars['Boolean']>;
  field_cooking_time?: Maybe<Scalars['Int']>;
  field_difficulty?: Maybe<Scalars['String']>;
  field_ingredients?: Maybe<Array<Maybe<Scalars['String']>>>;
  field_number_of_servings?: Maybe<Scalars['Int']>;
  field_preparation_time?: Maybe<Scalars['Int']>;
  field_recipe_instruction?: Maybe<Node__RecipeField_Recipe_Instruction>;
  field_summary?: Maybe<Node__RecipeField_Summary>;
};
// ...
export type Node__RecipeField_Recipe_Instruction = {
  value?: Maybe<Scalars['String']>;
  format?: Maybe<Scalars['String']>;
  processed?: Maybe<Scalars['String']>;
};

export type Node__RecipeField_Recipe_InstructionFilterInput = {
  value?: Maybe<StringQueryOperatorInput>;
  format?: Maybe<StringQueryOperatorInput>;
  processed?: Maybe<StringQueryOperatorInput>;
};

export type Node__RecipeField_Summary = {
  value?: Maybe<Scalars['String']>;
  format?: Maybe<Scalars['String']>;
  processed?: Maybe<Scalars['String']>;
};
// ...
export type Node__RecipePath = {
  alias?: Maybe<Scalars['String']>;
  pid?: Maybe<Scalars['Int']>;
  langcode?: Maybe<Scalars['String']>;
};
// ...
export type RecipeDetailsQuery = { translations: { nodes: Array<(
      Pick<Node__Recipe, 'langcode'>
      & { path?: Maybe<Pick<Node__RecipePath, 'alias'>> }
    )> }, content?: Maybe<(
    Pick<Node__Recipe, 'title'>
    & { field_summary?: Maybe<Pick<Node__RecipeField_Summary, 'value'>>, field_recipe_instruction?: Maybe<Pick<Node__RecipeField_Recipe_Instruction, 'processed'>> }
  )> };